fix(build): clear the two remaining -Wall -Wextra warnings
-Wnarrowing in engine_kv_cache_init.cpp (std::to_underlying int -> uint32_t
in braced init) and -Wunused-result in test_tool_stream_filter.cpp (GCC's
warn_unused_result attribute ignores a (void) cast; use std::ignore).

+++ src/runtime/engine_kv_cache_init.cpp
@@ -40,7 +40,7 @@ uint64_t Engine::model_fingerprint_() const {
     uint64_t h = 0xcbf29ce484222325ULL;
     const auto& c = model_->config();
     const uint32_t ids[] = {
-        std::to_underlying(c.arch),    static_cast<uint32_t>(c.n_layers),
+        static_cast<uint32_t>(std::to_underlying(c.arch)), static_cast<uint32_t>(c.n_layers),
         static_cast<uint32_t>(c.n_heads), static_cast<uint32_t>(c.n_kv_heads),
         static_cast<uint32_t>(c.d_model), static_cast<uint32_t>(c.d_ff),
         static_cast<uint32_t>(c.vocab_size), static_cast<uint32_t>(c.head_dim),
